MAINT: Move the `np.core.numerictypes` annotations to their own stub file
