+++ numpy/core/numerictypes.pyi
@@ -1,0 +1,26 @@
+from typing import TypeVar, Optional, Type, Union, Tuple, Sequence, overload, Any
+
+from numpy import generic, ndarray, dtype
+from numpy.typing import DtypeLike
+
+_DefaultType = TypeVar("_DefaultType")
+
+def maximum_sctype(t: DtypeLike) -> dtype: ...
+def issctype(rep: object) -> bool: ...
+@overload
+def obj2sctype(rep: object) -> Optional[generic]: ...
+@overload
+def obj2sctype(rep: object, default: None) -> Optional[generic]: ...
+@overload
+def obj2sctype(
+    rep: object, default: Type[_DefaultType]
+) -> Union[generic, Type[_DefaultType]]: ...
+def issubclass_(arg1: object, arg2: Union[object, Tuple[object, ...]]) -> bool: ...
+def issubsctype(
+    arg1: Union[ndarray, DtypeLike], arg2: Union[ndarray, DtypeLike]
+) -> bool: ...
+def issubdtype(arg1: DtypeLike, arg2: DtypeLike) -> bool: ...
+def sctype2char(sctype: object) -> str: ...
+def find_common_type(
+    array_types: Sequence[DtypeLike], scalar_types: Sequence[DtypeLike]
+) -> dtype: ...
